Update external.h
UnLoadLibrary essentially a dlclose() call on samsungbackend.so shared library causes the issue. Previously the workaround has been applied on driverWrapper.close() function call in MLperfDriverWrapper.java file. However, I think commenting this line makes it more specific to where the issue actually lies and the code is closer to the backend too.